Initial breaking out of common code
