+++ plugins/docker/attacher.go
@@ -29,8 +29,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"path/filepath"
-	"strings"
 	"time"
 
 	"github.com/fsouza/go-dockerclient"
@@ -39,9 +37,6 @@ import (
 	"github.com/pborman/uuid"
 )
 
-const (
-	SLEEP_BETWEEN_RECONNECT = 500 * time.Millisecond
-)
 
 type AttachManager struct {
 	hostname         string
@@ -55,22 +50,7 @@ type AttachManager struct {
 	fieldsFromLabels []string
 }
 
-// Return a properly configured Docker client
-func newDockerClient(certPath string, endpoint string) (DockerClient, error) {
-	var client DockerClient
-	var err error
 
-	if certPath == "" {
-		client, err = docker.NewClient(endpoint)
-	} else {
-		key := filepath.Join(certPath, "key.pem")
-		ca := filepath.Join(certPath, "ca.pem")
-		cert := filepath.Join(certPath, "cert.pem")
-		client, err = docker.NewTLSClient(endpoint, cert, key, ca)
-	}
-
-	return client, err
-}
 
 // Construct an AttachManager and set up the Docker Client
 func NewAttachManager(endpoint string, certPath string, nameFromEnv string,
@@ -90,34 +70,6 @@ func NewAttachManager(endpoint string, certPath string, nameFromEnv string,
 	}
 
 	return m, nil
-}
-
-// Handler to wrap functions with retry logic
-func withRetries(doWork func() error) error {
-	var err error
-
-	retrier, err := NewRetryHelper(RetryOptions{
-		MaxRetries: 10,
-		Delay:      "1s",
-		MaxJitter:  "250ms",
-	})
-	if err != nil {
-		return err
-	}
-
-	for {
-		err = doWork()
-		if err == nil {
-			return nil
-		}
-
-		// Sleep between retries, break if we're done
-		if e := retrier.Wait(); e != nil {
-			break
-		}
-	}
-
-	return err
 }
 
 // Attach to all running containers
@@ -219,42 +171,11 @@ func (m *AttachManager) handleDockerEvents(stopChan chan error) {
 	}
 }
 
-// Inspect the container and extract the env vars/labels we were told to keep
-func (m *AttachManager) extractFields(id string, client DockerClient) (map[string]string, error) {
-
-	container, err := client.InspectContainer(id)
-	if err != nil {
-		return nil, err
-	}
-	name := container.Name[1:] // Strip the leading slas
-	image := container.Config.Image
-
-	fields := getEnvVars(container, append(m.fieldsFromEnv, m.nameFromEnv))
-	if m.nameFromEnv != "" {
-		if alt_name, ok := fields[m.nameFromEnv]; ok && alt_name != "" {
-			name = alt_name
-		}
-	}
-	fields["ContainerID"] = id
-	fields["ContainerName"] = name
-	fields["ContainerImage"] = image
-
-	// NOTE! Anything that is a duplicate key will be overridden with the value
-	// that is in the label, not the environment
-	for _, key := range m.fieldsFromLabels {
-		if value, ok := container.Config.Labels[key]; ok {
-			fields[key] = value
-		}
-	}
-
-	return fields, nil
-}
-
 // Attach to the output of a single running container
 func (m *AttachManager) attach(id string, client DockerClient) error {
 	m.ir.LogMessage(fmt.Sprintf("Attaching container: %s", id))
 
-	fields, err := m.extractFields(id, client)
+	fields, err := extractFields(id, client, m.fieldsFromLabels, m.fieldsFromEnv, m.nameFromEnv)
 	if err != nil {
 		return err
 	}
@@ -352,19 +273,3 @@ func (m *AttachManager) handleOneStream(name string, in io.Reader, fields map[st
 	m.ir.LogMessage(fmt.Sprintf("Disconnecting %s stream from %s", name, containerId))
 }
 
-// Process the env vars and capture the ones we want
-func getEnvVars(container *docker.Container, keys []string) map[string]string {
-	vars := make(map[string]string)
-	for _, value := range container.Config.Env {
-		valueParts := strings.SplitN(value, "=", 2)
-		if len(valueParts) == 2 {
-			for _, key := range keys {
-				if key != "" && valueParts[0] == key {
-					vars[valueParts[0]] = valueParts[1]
-					break
-				}
-			}
-		}
-	}
-	return vars
-}
